more typing updates

+++ supermarq-benchmarks/supermarq/plotting.py
@@ -372,8 +372,8 @@ def radar_factory(num_vars: int) -> npt.NDArray[np.float64]:
     Returns:
         A list of evenly spaced angles.
     """
-    # calculate evenly-spaced axis angles
-    theta = np.linspace(0, 2 * np.pi, num_vars, endpoint=False)
+    # Calculate evenly-spaced axis angles
+    theta = np.linspace(0, 2 * np.pi, num_vars, endpoint=False, dtype=np.float64)
 
     class RadarAxes(RadarAxesMeta):
         """A helper class that sets the shape of the feature plot"""
